Algorithm cleanup.  - Algos module renamed to JWA  - Standard HMAC algorithms provided explicitly by OpenSSL (#550)  - Prepare to remove support for the HS512256 algorithm provided by RbNaCl (#549)
