+++ lib/jwt/jwa/rsa.rb
@@ -1,14 +1,16 @@
 # frozen_string_literal: true
 
 module JWT
-  module Algos
+  module JWA
     module Rsa
       module_function
 
       SUPPORTED = %w[RS256 RS384 RS512].freeze
 
       def sign(algorithm, msg, key)
-        raise EncodeError, "The given key is a #{key.class}. It has to be an OpenSSL::PKey::RSA instance." if key.instance_of?(String)
+        unless key.is_a?(OpenSSL::PKey::RSA)
+          raise EncodeError, "The given key is a #{key.class}. It has to be an OpenSSL::PKey::RSA instance"
+        end
 
         key.sign(OpenSSL::Digest.new(algorithm.sub('RS', 'sha')), msg)
       end
