+++ spec/jwt/jwa/ecdsa_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe JWT::Algos::Ecdsa do
+RSpec.describe JWT::JWA::Ecdsa do
   describe '.curve_by_name' do
     subject { described_class.curve_by_name(curve_name) }
