+++ spec/jwt/jwt_spec.rb
@@ -778,7 +778,7 @@ RSpec.describe JWT do
     let(:token) { JWT.encode(payload, 'secret', 'HS256') }
 
     it 'starts trying with the algorithm referred in the header' do
-      expect(JWT::Algos::Rsa).not_to receive(:verify)
+      expect(JWT::JWA::Rsa).not_to receive(:verify)
       JWT.decode(token, 'secret', true, algorithm: ['RS512', 'HS256'])
     end
   end
